+++ parse.c
@@ -133,9 +133,9 @@ static vString* determineInterpreter (const char* const cmd)
     do
     {
 	vStringClear (interpreter);
-	for ( ;  isspace (*p)  ;  ++p)
+	for ( ;  isspace ((int) *p)  ;  ++p)
 	    ; /* no-op */
-	for ( ;  *p != '\0'  &&  ! isspace (*p)  ;  ++p)
+	for ( ;  *p != '\0'  &&  ! isspace ((int) *p)  ;  ++p)
 	    vStringPut (interpreter, (int) *p);
 	vStringTerminate (interpreter);
     } while (strcmp (vStringValue (interpreter), "env") == 0);
